Move gorilla to subpackage
